REFACTOR nav to component
