+++ src/App.css
@@ -1,8 +1,13 @@
 body {
   font-family: "Noto Sans JP", sans-serif; }
 
-.hero-head {
-  padding-top: 3rem; }
+nav {
+  padding-top: 2rem;
+  padding-bottom: 1.5rem; }
+
+.button:hover {
+  color: #fff;
+  opacity: 0.8; }
 
 .container {
   max-width: 1024px; }
@@ -95,9 +100,6 @@ body {
     margin-left: 15px;
     font-size: 0.9rem;
     line-height: 1; }
-  .stack .stack-link:hover {
-    color: #fff;
-    opacity: 0.8; }
   .stack .columns .column:first-of-type {
     padding-left: 30px; }
   .stack .columns .column {
